requirements.txt
numpy
opencv-python
scipy
pandas
tqdm
easydict
pillow
exifread